requirements.txt
# Core data processing
numpy==1.24.3
pandas==2.0.3

# HTTP requests for APIs
requests==2.31.0

# Environment management
python-dotenv==1.0.0

# For future NLP tasks
scikit-learn==1.3.0

# Jupyter for experimentation (optional)
jupyter==1.0.0